+++ test.sh
@@ -29,6 +29,8 @@ Rao_RepH_GM12878_InsulationScore.txt.multires.db
 hic-resolutions.cool
 wgEncodeCaltechRnaSeqHuvecR1x75dTh1014IlnaPlusSignalRep2.bigWig
 wgEncodeCaltechRnaSeqHuvecR1x75dTh1014IlnaPlusSignalRep2.hitile
+chr21.KL.bed.multires.mv5
+sample.bed.multires.mv5
 END
 )