+++ tilesets/views.py
@@ -21,6 +21,13 @@ import itertools as it
 
 import tilesets.chromsizes as tcs
 import tilesets.generate_tiles as tgt
+import tilesets.multivec_tiles as tmt
+
+import hgtiles.cooler as hgco
+import hgtiles.bigwig as hgbi
+import hgtiles.multivec as hgmu
+
+import tilesets.chromsizes as tcs
 import tilesets.models as tm
 import tilesets.permissions as tsp
 import tilesets.serializers as tss
@@ -109,28 +116,22 @@ def available_chrom_sizes(request):
 @authentication_classes((CsrfExemptSessionAuthentication, BasicAuthentication))
 def sizes(request):
     '''Return chromosome sizes.
-
     Retrieves the chromSiyes.tsv and either retrieves it as is or converts it
     to a JSON format.
-
     Args:
         request: HTTP GET request object. The request can feature the following
             queries:
-
             id: id of the stored chromSizes [e.g.: hg19 or mm9]
-            type: return data format [csv, tsv, or json]
+            type: return data format [tsv or json]
             cum: return cumulative size or offset [0 or 1]
-
     Returns:
         A HTTP text or JSON response depending on the GET request.
-
         A text response looks like this:
         ```
         chr1    1
         chr2    2
         ...
         ```
-
         A JSON response looks like this:
         ```
         {
@@ -167,7 +168,7 @@ def sizes(request):
     try:
         chrom_sizes = tm.Tileset.objects.get(uuid=uuid)
     except Exception as e:
-        logger.error(e)
+        logger.exception(e)
         err_msg = 'Oh lord! ChromSizes for %s not found. 😬' % uuid
         err_status = 404
 
@@ -179,16 +180,29 @@ def sizes(request):
     # Try to load the chromosome sizes and return them as a list of
     # (name, size) tuples
     try:
-        if chrom_sizes.filetype == 'cooler':
+        if tgt.get_tileset_filetype(chrom_sizes) == 'bigwig':
+            data = hgbi.chromsizes(
+                tut.get_datapath(chrom_sizes.datafile.url)
+            )
+        elif tgt.get_tileset_filetype(chrom_sizes) == 'cooler':
             data = tcs.get_cooler_chromsizes(
                 tut.get_datapath(chrom_sizes.datafile.url)
             )
-        else:
+        elif tgt.get_tileset_filetype(chrom_sizes) == 'chromsizes-tsv':
             data = tcs.get_tsv_chromsizes(
                 tut.get_datapath(chrom_sizes.datafile.url)
             )
+        elif tgt.get_tileset_filetype(chrom_sizes) == 'multivec':
+            data = tcs.get_multivec_chromsizes(
+                tut.get_datapath(chrom_sizes.datafile.url)
+            )
+        else:
+            data = ''
+
     except Exception as ex:
+        logger.exception(ex)
         err_msg = str(ex)
+        print('err_msg:', err_msg)
         err_status = 500
 
         if is_json:
@@ -232,7 +246,7 @@ def sizes(request):
 
             data = json_out
     except Exception as e:
-        logger.error(e)
+        logger.exception(e)
         err_msg = 'THIS IS AN OUTRAGE!!!1! Something failed. 😡'
         err_status = 500
 
@@ -359,7 +373,7 @@ def add_transform_type(tile_id):
     tileset_uuid = tile_id_parts[0]
     tile_position = tile_id_parts[1:4]
 
-    transform_type = tgt.get_transform_type(tile_id)
+    transform_type = hgco.get_transform_type(tile_id)
     new_tile_id = ".".join([tileset_uuid] + tile_position + [transform_type])
     return new_tile_id
 
@@ -512,6 +526,28 @@ def tileset_info(request):
     queryset = tm.Tileset.objects.all()
     tileset_uuids = request.GET.getlist("d")
     tileset_infos = {}
+
+    chromsizes_error = None
+
+    if 'cs' in request.GET:
+        # we need to call a different server to get the tiles
+        if 'ci' not in request.GET.getlist:
+            chromsizes_error = 'cs param present without ci'
+
+        # call the request server and get the chromsizes
+        pass
+    else:
+        if 'ci' in request.GET:
+            try:
+                chromsizes = tm.Tileset.objects.get(uuid=request.GET['ci'])
+                data = tcs.chromsizes_array_to_series(
+                        tcs.get_tsv_chromsizes(
+                            tut.get_datapath(chromsizes.datafile.url))
+                        )
+                print('data:', data)
+            except Exception as ex:
+                pass
+
     for tileset_uuid in tileset_uuids:
         tileset_object = queryset.filter(uuid=tileset_uuid).first()
 
@@ -559,7 +595,7 @@ def tileset_info(request):
                 tileset_object
             )
         elif tileset_object.filetype == 'multivec':
-            tileset_infos[tileset_uuid] = tgt.generate_multivec_tileset_info(
+            tileset_infos[tileset_uuid] = hgmu.tileset_info(
                     tut.get_datapath(tileset_object.datafile.url))
         elif tileset_object.filetype == "elastic_search":
             response = urllib.urlopen(
@@ -585,17 +621,13 @@ def tileset_info(request):
                 tut.get_datapath(tileset_object.datafile.url)
             )
         elif tileset_object.filetype == 'cooler':
-            dsetname = tut.get_datapath(queryset.filter(
-                uuid=tileset_uuid
-            ).first().datafile.url)
-
-            if dsetname not in tgt.mats:
-                tgt.make_mats(dsetname)
-            tileset_infos[tileset_uuid] = tgt.mats[dsetname][1]
+            tileset_infos[tileset_uuid] = hgco.tileset_info(
+                    tut.get_datapath(tileset_object.datafile.url)
+            )
         else:
             # Unknown filetype
             tileset_infos[tileset_uuid] = {
-                'message': 'Unknown filetype ' + tileset_object.filetype
+                'error': 'Unknown filetype ' + tileset_object.filetype
             }
 
         tileset_infos[tileset_uuid]['name'] = tileset_object.name